Refactor: MailPuzzleResult package 수정
